chore: move api setup into lib dir
