+++ app/api/auth/[...nextauth]/route.ts
@@ -1,142 +1,5 @@
-import db from '@/lib/prisma';
-import { PrismaAdapter } from '@next-auth/prisma-adapter';
-import { compare } from 'bcrypt';
-import NextAuth, { AuthOptions, TokenSet } from 'next-auth';
-import CredentialsProvider from 'next-auth/providers/credentials';
-import GithubProvider from 'next-auth/providers/github';
-import GoogleProvider from 'next-auth/providers/google';
-
-const GOOGLE_CLIENT_ID = process.env.GOOGLE_CLIENT_ID || '';
-const GOOGLE_CLIENT_SECRET = process.env.GOOGLE_CLIENT_SECRET || '';
-
-export const authOptions: AuthOptions = {
-  adapter: PrismaAdapter(db),
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_ID || '',
-      clientSecret: process.env.GITHUB_SECRET || '',
-    }),
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID || '',
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET || '',
-      authorization: {
-        params: {
-          access_type: 'offline',
-          prompt: 'consent',
-          response_type: 'code',
-        },
-      },
-    }),
-    CredentialsProvider({
-      // The name to display on the sign in form (e.g. "Sign in with...")
-      name: 'Credentials',
-      // `credentials` is used to generate a form on the sign in page.
-      // You can specify which fields should be submitted, by adding keys to the `credentials` object.
-      // e.g. domain, username, password, 2FA token, etc.
-      // You can pass any HTML attribute to the <input> tag through the object.
-      credentials: {
-        email: {
-          label: 'email',
-          type: 'text',
-          placeholder: 'jsmith@example.com',
-        },
-        password: { label: 'password', type: 'password' },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password)
-          throw new TypeError(
-            'Invalid Credentials: Email and password are required'
-          );
-
-        const user = await db.user.findUnique({
-          where: {
-            email: credentials.email.toLowerCase(),
-          },
-        });
-        if (!user || !user?.password)
-          throw new ReferenceError(
-            'Invalid Credentials: This user does not exist'
-          );
-
-        const matches = await compare(credentials.password, user.password);
-        if (!matches) {
-          throw new TypeError(
-            'Invalid Credentials: Please enter a valid password'
-          );
-        }
-
-        return user;
-      },
-    }),
-  ],
-  callbacks: {
-    async session({ session, user }) {
-      const [google] = await db.account.findMany({
-        where: { userId: user.id, provider: 'google' },
-      });
-
-      if (google?.expires_at && google?.expires_at * 1000 < Date.now()) {
-        // If the access token has expired, try to refresh it
-        try {
-          // https://accounts.google.com/.well-known/openid-configuration
-          // We need the `token_endpoint`.
-          const response = await fetch('https://oauth2.googleapis.com/token', {
-            headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-            body: new URLSearchParams({
-              client_id: GOOGLE_CLIENT_ID,
-              client_secret: GOOGLE_CLIENT_SECRET,
-              grant_type: 'refresh_token',
-              refresh_token: google.refresh_token || '',
-            }),
-            method: 'POST',
-          });
-
-          const tokens: TokenSet = await response.json();
-          if (!response.ok) throw tokens;
-
-          await db.account.update({
-            data: {
-              access_token: tokens.access_token,
-              expires_at: Math.floor(
-                Date.now() / 1000 + Number(tokens.expires_in)
-              ),
-              refresh_token: tokens.refresh_token ?? google.refresh_token,
-            },
-            where: {
-              provider_providerAccountId: {
-                provider: 'google',
-                providerAccountId: google.providerAccountId,
-              },
-            },
-          });
-        } catch (error) {
-          console.error('Error refreshing access token', error);
-          // The error property will be used client-side to handle the refresh token error
-          session.error = 'RefreshAccessTokenError';
-        }
-      }
-      return session;
-    },
-  },
-
-  pages: {
-    signIn: '/',
-    // signIn: '/auth/signin',
-    // signOut: '/auth/signout',
-    // error: '/auth/error', // Error code passed in query string as ?error=
-    // verifyRequest: '/auth/verify-request', // (used for check email message)
-    // newUser: '/auth/new-user' // New users will be directed here on first sign in (leave the property out if not of interest)
-  },
-
-  session: {
-    strategy: 'database',
-    maxAge: 30 * 24 * 60 * 60, // 30 days
-    updateAge: 1 * 24 * 60 * 60, // 1 hours
-  },
-
-  secret: process.env.NEXTAUTH_SECRET,
-  debug: process.env.NODE_ENV === 'development',
-};
+import authOptions from '@/lib/api/auth-options';
+import NextAuth from 'next-auth';
 
 const handler = NextAuth(authOptions);
