+++ lib/data/get-session.ts
@@ -1,5 +1,5 @@
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../api/auth/[...nextauth]/route';
+import authOptions from '../api/auth-options';
 
 export async function getSession() {
   return await getServerSession(authOptions);
