+++ lib/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './use-form';
 export * from './use-media';
 export * from './use-modals';
-export * from './use-queries';
 export * from './use-theme';
